// src/enums/token-type.ts
enum TokenType {
  //Symbols
  LEFT_PAREN = 'LEFT_PAREN', 
  RIGHT_PAREN = 'RIGHT_PAREN',
  COMMA = 'COMMA',
  DOT = 'DOT',
  MINUS = 'MINUS',
  PLUS = 'PLUS',
  SLASH = 'SLASH',
  STAR = 'STAR',
  EXPONENTIATION = 'EXPONENTIATION',
  //1 or 2 character tokens
  BANG = 'BANG', 
  BANG_EQUAL = 'BANG_EQUAL',
  EQUAL = 'EQUAL', 
  EQUAL_EQUAL = 'EQUAL_EQUAL',
  GREATER = 'GREATER', 
  GREATER_EQUAL = 'GREATER_EQUAL',
  LESS = 'LESS',
  LESS_EQUAL = 'LESS_EQUAL',
  GREATER_LESS_NOT_EQUAL = 'GREATER_LESS_NOT_EQUAL',
  AMBERSAND = 'AMBERSAND',
  AMBERSAND_AMBERSAND = 'AMBERSAND_AMBERSAND',
  PIPE_PIPE = 'PIPE_PIPE',
  //Literals
  IDENTIFIER = 'IDENTIFIER',
  STRING = 'STRING',
  NUMBER = 'NUMBER',
  //Keywords
  TRUE = 'TRUE',
  FALSE = 'FALSE',
  FUNC_ABS = 'FUNC_ABS',
  FUNC_ADDMONTHS = 'FUNC_ADDMONTHS',
  FUNC_AND = 'FUNC_AND',
  FUNC_BEGINS = 'FUNC_BEGINS',
  FUNC_BLANKVALUE = 'FUNC_BLANKVALUE',
  FUNC_BR = 'FUNC_BR',
  FUNC_CASE = 'FUNC_CASE',
  FUNC_CASESAFEID = 'FUNC_CASESAFEID',
  FUNC_CEILING = 'FUNC_CEILING',
  FUNC_CONTAINS = 'FUNC_CONTAINS',
  FUNC_CURRENCYRATE = 'FUNC_CURRENCYRATE',
  FUNC_DATE = 'FUNC_DATE',
  FUNC_DATEVALUE = 'FUNC_DATEVALUE',
  FUNC_DATETIMEVALUE = 'FUNC_DATETIMEVALUE',
  FUNC_DAY = 'FUNC_DAY',
  FUNC_DISTANCE = 'FUNC_DISTANCE',
  FUNC_EXP = 'FUNC_EXP',
  FUNC_FIND = 'FUNC_FIND',
  FUNC_FLOOR = 'FUNC_FLOOR',
  FUNC_GEOLOCATION = 'FUNC_GEOLOCATION',
  FUNC_GETRECORDIDS = 'FUNC_GETRECORDIDS',
  FUNC_GETSESSIONID = 'FUNC_GETSESSIONID',
  FUNC_HOUR = 'FUNC_HOUR',
  FUNC_HTMLENCODE = 'FUNC_HTMLENCODE',
  FUNC_HYPERLINK = 'FUNC_HYPERLINK',
  FUNC_IF = 'FUNC_IF',
  FUNC_IMAGE = 'FUNC_IMAGE',
  FUNC_IMAGEPROXYURL = 'FUNC_IMAGEPROXYURL',
  FUNC_INCLUDE = 'FUNC_INCLUDE',
  FUNC_INCLUDES = 'FUNC_INCLUDES',
  FUNC_ISBLANK = 'FUNC_ISBLANK',
  FUNC_ISCHANGED = 'FUNC_ISCHANGED',
  FUNC_ISCLONE = 'FUNC_ISCLONE',
  FUNC_ISNEW = 'FUNC_ISNEW',
  FUNC_ISNULL = 'FUNC_ISNULL',
  FUNC_ISNUMBER = 'FUNC_ISNUMBER',
  FUNC_ISPICKVAL = 'FUNC_ISPICKVAL',
  FUNC_JSENCODE = 'FUNC_JSENCODE',
  FUNC_JSINHTMLENCODE = 'FUNC_JSINHTMLENCODE',
  FUNC_JUNCTIONIDLIST = 'FUNC_JUNCTIONIDLIST',
  FUNC_LEFT = 'FUNC_LEFT',
  FUNC_LEN = 'FUNC_LEN',
  FUNC_LINKTO = 'FUNC_LINKTO',
  FUNC_LN = 'FUNC_LN',
  FUNC_LOG = 'FUNC_LOG',
  FUNC_LOWER = 'FUNC_LOWER',
  FUNC_LPAD = 'FUNC_LPAD',
  FUNC_MAX = 'FUNC_MAX',
  FUNC_MCEILING = 'FUNC_MCEILING',
  FUNC_MFLOOR = 'FUNC_MFLOOR',
  FUNC_MID = 'FUNC_MID',
  FUNC_MILLISECOND = 'FUNC_MILLISECOND',
  FUNC_MIN = 'FUNC_MIN',
  FUNC_MINUTE = 'FUNC_MINUTE',
  FUNC_MOD = 'FUNC_MOD',
  FUNC_MONTH = 'FUNC_MONTH',
  FUNC_NOT = 'FUNC_NOT',
  FUNC_NOW = 'FUNC_NOW',
  FUNC_NULLVALUE = 'FUNC_NULLVALUE',
  FUNC_OR = 'FUNC_OR',
  FUNC_PARENTGROUPVAL = 'FUNC_PARENTGROUPVAL',
  FUNC_PREDICT = 'FUNC_PREDICT',
  FUNC_PREVGROUPVAL = 'FUNC_PREVGROUPVAL',
  FUNC_PRIORVALUE = 'FUNC_PRIORVALUE',
  FUNC_REGEX = 'FUNC_REGEX',
  FUNC_REQUIRESCRIPT = 'FUNC_REQUIRESCRIPT',
  FUNC_REVERSE = 'FUNC_REVERSE',
  FUNC_RIGHT = 'FUNC_RIGHT',
  FUNC_ROUND = 'FUNC_ROUND',
  FUNC_RPAD = 'FUNC_RPAD',
  FUNC_SECOND = 'FUNC_SECOND',
  FUNC_SQRT = 'FUNC_SQRT',
  FUNC_SUBSTITUTE = 'FUNC_SUBSTITUTE',
  FUNC_TEXT = 'FUNC_TEXT',
  FUNC_TIMENOW = 'FUNC_TIMENOW',
  FUNC_TIMEVALUE = 'FUNC_TIMEVALUE',
  FUNC_TODAY = 'FUNC_TODAY',
  FUNC_TRIM = 'FUNC_TRIM',
  FUNC_UPPER = 'FUNC_UPPER',
  FUNC_URLENCODE = 'FUNC_URLENCODE',
  FUNC_URLFOR = 'FUNC_URLFOR',
  FUNC_VALUE = 'FUNC_VALUE',
  FUNC_VLOOKUP = 'FUNC_VLOOKUP',
  FUNC_WEEKDAY = 'FUNC_WEEKDAY',
  FUNC_YEAR = 'FUNC_YEAR',
  //System
  EOF = 'EOF'
}
export default TokenType;